requirements.txt
behave==1.2.5
defusedxml==0.5.0
Django==1.10.5
djangorestframework==3.6.3
djangorestframework-xml==1.3.0
enum34==1.1.6
parse==1.8.0
parse-type==0.3.4
pytz==2017.2
requests==2.14.2
selenium==3.4.2
six==1.10.0
splinter==0.7.5
